requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.1
backoff==2.2.1
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.402
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dataclasses-json==0.5.14
Deprecated==1.2.14
fastapi==0.99.1
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
langchain==0.0.279
langchain-experimental==0.0.13
langsmith==0.0.33
Lazify==0.4.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nodeenv==1.8.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
packaging==23.1
prisma==0.10.0
protobuf==4.24.2
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-engineio==4.6.1
python-graphql-client==0.4.3
python-socketio==5.8.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
uptrace==1.19.0
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
